+++ config/config.json
@@ -1,14 +1,14 @@
 {
     "development": {
       "username": "root",
-      "password": "password",
+      "password": "",
       "database": "Message",
       "host": "127.0.0.1",
       "dialect": "mysql"
     },
     "test": {
       "username": "root",
-      "password": "password",
+      "password": "",
       "database": "Message",
       "host": "127.0.0.1",
       "dialect": "mysql"
