+++ package.json
@@ -4,6 +4,7 @@
   "description": "",
   "main": "node server.js",
   "scripts": {
+    "start": "node server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
