+++ server.js
@@ -5,17 +5,18 @@ const db = require("./models");
 
 // Creating express app and configuring middleware needed for authentication
 const app = express();
-
+app.use(express.urlencoded({ extended: true }));
+app.use(express.json());
+app.use(express.static(__dirname + '/public'));
 var exphbs = require("express-handlebars");
 app.engine("handlebars", exphbs({ defaultLayout: "main"}));
 app.set("view engine", "handlebars");
-app.use(express.static(__dirname + '/public'));
-app.use(express.urlencoded({ extended: true }));
-app.use(express.json());
+
 //app.use(express.static("public"));
 
 require("./routes/html-routes.js")(app)
 require("./routes/api-routes.js")(app);
+
 
 // Syncing our database and logging a message to the user upon success
 db.sequelize.sync().then(() => {
